refactor: 카드 이미지 컴포넌트 분리
