+++ src/screens/home/HomeScreen.tsx
@@ -8,7 +8,6 @@ import {
   Dimensions,
   NativeSyntheticEvent,
   NativeScrollEvent,
-  ImageLoadEventData,
 } from 'react-native';
 import {
   Container,
@@ -21,7 +20,6 @@ import {
   Product,
   ProductName,
   ProductInfo,
-  CardImg,
   CardContainer,
   CardInfo,
   Button,
@@ -43,6 +41,7 @@ import {
   CardBenefitSmallText,
 } from './HomeScreen.styled';
 import axios from 'axios';
+import CardImage from '../../components/CardImage/CardImage';
 
 const { width: windowWidth } = Dimensions.get('window');
 
@@ -88,38 +87,6 @@ const Pagination: React.FC<{ length: number; currentIndex: number }> =
       ))}
     </PaginationContainer>
   ));
-
-  interface CardImageProps {
-    uri: string;
-    ImgHeight: number;
-    ImgWidth: number;
-  }
-
-  const CardImage: React.FC<CardImageProps> = ({ uri, ImgHeight, ImgWidth }) => {
-    const [shouldRotate, setShouldRotate] = useState(false);
-  
-    const handleImageLoad = (event: NativeSyntheticEvent<ImageLoadEventData>) => {
-      const { width, height } = event.nativeEvent.source;
-
-      // 이미지의 가로가 세로보다 길 경우 회전 플래그 설정
-      if (width > height) {
-        setShouldRotate(true);
-      }
-    };
-  
-    return (
-      <Image
-        source={{ uri }}
-        style={{
-          width: shouldRotate ? ImgHeight : ImgWidth, // 회전 후 가로 길이
-          height: shouldRotate ? ImgWidth : ImgHeight, // 회전 후 세로 길이
-          transform: shouldRotate ? [{ rotate: '90deg' }] : [], // 회전 조건 추가
-          objectFit: "contain"
-        }}
-        onLoad={handleImageLoad}
-      />
-    );
-  };
 
 const CardList = React.memo(
   ({ SelectedCardList }: { SelectedCardList: any[] }) => (
